Create FunctionsManager to share config scope with functions
