+++ interpreter/config/index.ts
@@ -1,2 +1,5 @@
+export * from "./config";
 export * from "./managers/color/manager";
 export * from "./managers/color/schema";
+export * from "./managers/functions";
+export * from "./managers/unit/manager";
