+++ interpreter/config/managers/functions/errors.ts
@@ -1,0 +1,5 @@
+export enum FunctionsManagerError {
+  INVALID_FUNCTION_NAME = "INVALID_FUNCTION_NAME",
+  FUNCTION_NOT_FOUND = "FUNCTION_NOT_FOUND",
+  INVALID_ARGUMENTS = "INVALID_ARGUMENTS",
+}
