+++ interpreter/config/managers/functions/index.ts
@@ -1,0 +1,3 @@
+export { FunctionsManagerError } from "./errors";
+export { FunctionsManager } from "./manager";
+export type { FunctionSpecification } from "./schema";
